+++ CS_project_ML/MyData.cpp
@@ -2,7 +2,7 @@
 
 MyData::MyData() {}
 
-MyData::MyData(int num, vector<float> features, int label) {
+MyData::MyData(int num, vector<double> features, int label) {
 	this->num = num;
 	this->features = features;
 	this->label = label;
